tests: add vue router to playground
